Moved   from unit to integration tests
