+++ test/integration/test_options.py
@@ -98,3 +98,25 @@ class TestIntegrationOptions(IBMIntegrationTestCase):
                         job2.result()
                     # TODO: Re-enable when ntc-1651 is fixed
                     # self.assertIn("TranspilerError", err.exception.message)
+
+    def test_optimization_level(self):
+        """Test various definitions for optimization_level."""
+
+        backend = "ibmq_qasm_simulator"
+        noise_model = NoiseModel.from_backend(FakeManila())
+        default_options = Options()
+        noisy_options = Options()
+        noisy_options.simulator.noise_model = noise_model
+        primitives = [Sampler, Estimator]
+        for cls in primitives:
+            cls_no_noise = cls(session=backend, options=default_options)
+            self.assertTrue(cls_no_noise.options.optimization_level == 1)
+
+            cls_with_noise = cls(session=backend, options=noisy_options)
+            self.assertTrue(cls_with_noise.options.optimization_level == 3)
+
+            user_given_options = Options()
+            for opt_level in [0, 1, 2, 3, 99]:
+                user_given_options.optimization_level = opt_level
+                cls_default = cls(session=backend, options=user_given_options)
+                self.assertTrue(cls_default.options.optimization_level == opt_level)
